Began splitting test_import.py
